+++ NetworkClient/src/ru/geekbrains/java2/client/controller/AuthEvent.java
@@ -1,0 +1,6 @@
+package ru.geekbrains.java2.client.controller;
+
+@FunctionalInterface
+public interface AuthEvent {
+    void authIsSuccessful(String nickname);
+}
