+++ NetworkClient/src/ru/geekbrains/java2/client/controller/MessageHandler.java
@@ -1,0 +1,6 @@
+package ru.geekbrains.java2.client.controller;
+
+@FunctionalInterface
+public interface MessageHandler {
+    void handle(String message);
+}
