+++ NetworkClientServer/src/ru/geekbrains/java2/client/CommandType.java
@@ -1,0 +1,12 @@
+package ru.geekbrains.java2.client;
+
+public enum CommandType {
+    AUTH,
+    AUTH_ERROR,
+    PRIVATE_MESSAGE,
+    BROADCAST_MESSAGE,
+    MESSAGE,
+    UPDATE_USERS_LIST,
+    ERROR,
+    END,
+}
