+++ NetworkClientServer/src/ru/geekbrains/java2/client/command/UpdateUsersListCommand.java
@@ -1,0 +1,17 @@
+package ru.geekbrains.java2.client.command;
+
+import java.io.Serializable;
+import java.util.List;
+
+public class UpdateUsersListCommand implements Serializable {
+
+    private final List<String> users;
+
+    public UpdateUsersListCommand(List<String> users) {
+        this.users = users;
+    }
+
+    public List<String> getUsers() {
+        return users;
+    }
+}
