+++ NetworkServer/src/ru/geekbrains/java2/server/auth/AuthService.java
@@ -1,0 +1,10 @@
+package ru.geekbrains.java2.server.auth;
+
+public interface AuthService {
+
+    String getUsernameByLoginAndPassword(String login, String password);
+
+    void start();
+    void stop();
+
+}
