--- src/messenger/client/NetworkClient.java
@@ -1,17 +1,0 @@
-package messenger.client;
-
-import messenger.client.controller.ClientController;
-
-import java.io.IOException;
-
-public class NetworkClient {
-
-    public static void main(String[] args) {
-        try {
-            ClientController clientController = new ClientController("localhost", 8190);
-            clientController.runApplication();
-        } catch (IOException e) {
-            System.err.println("Failed to connect to server! Please, check you network settings");
-        }
-    }
-}
