--- src/messenger/client/client_server/CommandType.java
@@ -1,12 +1,0 @@
-package messenger.client.client_server;
-
-public enum CommandType {
-    AUTH,
-    AUTH_ERROR,
-    PRIVATE_MESSAGE,
-    BROADCAST_MESSAGE,
-    MESSAGE,
-    UPDATE_USERS_LIST,
-    ERROR,
-    END,
-}
