--- src/messenger/client/client_server/command/BroadcastMessageCommand.java
@@ -1,16 +1,0 @@
-package messenger.client.client_server.command;
-
-import java.io.Serializable;
-
-public class BroadcastMessageCommand implements Serializable {
-
-    private final String message;
-
-    public BroadcastMessageCommand(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
