--- src/messenger/client/client_server/command/PrivateMessageCommand.java
@@ -1,22 +1,0 @@
-package messenger.client.client_server.command;
-
-import java.io.Serializable;
-
-public class PrivateMessageCommand implements Serializable {
-
-    private final String receiver;
-    private final String message;
-
-    public PrivateMessageCommand(String receiver, String message) {
-        this.receiver = receiver;
-        this.message = message;
-    }
-
-    public String getReceiver() {
-        return receiver;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
