--- src/messenger/client/client_server/command/UpdateUsersListCommand.java
@@ -1,17 +1,0 @@
-package messenger.client.client_server.command;
-
-import java.io.Serializable;
-import java.util.List;
-
-public class UpdateUsersListCommand implements Serializable {
-
-    private final List<String> users;
-
-    public UpdateUsersListCommand(List<String> users) {
-        this.users = users;
-    }
-
-    public List<String> getUsers() {
-        return users;
-    }
-}
