--- src/messenger/client/controller/AuthEvent.java
@@ -1,6 +1,0 @@
-package messenger.client.controller;
-
-@FunctionalInterface
-public interface AuthEvent {
-    void authIsSuccessful(String nickname);
-}
