--- src/messenger/client/controller/MessageHandler.java
@@ -1,6 +1,0 @@
-package messenger.client.controller;
-
-@FunctionalInterface
-public interface MessageHandler {
-    void handle(String message);
-}
