--- src/messenger/client/server/auth/AuthService.java
@@ -1,10 +1,0 @@
-package messenger.client.server.auth;
-
-public interface AuthService {
-
-    String getUsernameByLoginAndPassword(String login, String password);
-
-    void start();
-    void stop();
-
-}
