--- src/messenger/client/view/ClientChat.form
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="homework.client.view.ClientChat">
-  <grid id="9b773" binding="mainPanel" default-binding="true" layout-manager="BorderLayout" hgap="0" vgap="0">
-    <constraints>
-      <xy x="101" y="84" width="640" height="480"/>
-    </constraints>
-    <properties>
-      <minimumSize width="640" height="480"/>
-    </properties>
-    <border type="none"/>
-    <children>
-      <grid id="f810b" layout-manager="GridLayoutManager" row-count="2" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-        <margin top="0" left="0" bottom="0" right="0"/>
-        <constraints border-constraint="Center"/>
-        <properties/>
-        <border type="etched"/>
-        <children>
-          <component id="e4b50" class="javax.swing.JTextArea" binding="chatText">
-            <constraints>
-              <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="6" anchor="0" fill="3" indent="0" use-parent-layout="false">
-                <preferred-size width="150" height="50"/>
-              </grid>
-            </constraints>
-            <properties/>
-          </component>
-          <grid id="879d6" layout-manager="GridLayoutManager" row-count="1" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-            <margin top="0" left="0" bottom="0" right="0"/>
-            <constraints>
-              <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
-            </constraints>
-            <properties/>
-            <border type="none"/>
-            <children>
-              <component id="e781a" class="javax.swing.JTextField" binding="messageTextField" default-binding="true">
-                <constraints>
-                  <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="3" indent="0" use-parent-layout="false">
-                    <preferred-size width="150" height="-1"/>
-                  </grid>
-                </constraints>
-                <properties>
-                  <margin top="3" left="8" bottom="3" right="8"/>
-                </properties>
-              </component>
-              <component id="af4d2" class="javax.swing.JButton" binding="sendButton" default-binding="true">
-                <constraints>
-                  <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
-                </constraints>
-                <properties>
-                  <text value="Send"/>
-                </properties>
-              </component>
-            </children>
-          </grid>
-        </children>
-      </grid>
-      <grid id="2e212" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-        <margin top="0" left="0" bottom="0" right="0"/>
-        <constraints border-constraint="West"/>
-        <properties>
-          <minimumSize width="50" height="-1"/>
-          <preferredSize width="150" height="-1"/>
-        </properties>
-        <border type="none"/>
-        <children>
-          <component id="a3c29" class="javax.swing.JList" binding="usersList" default-binding="true">
-            <constraints>
-              <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="2" anchor="0" fill="3" indent="0" use-parent-layout="false">
-                <preferred-size width="150" height="50"/>
-              </grid>
-            </constraints>
-            <properties>
-              <background color="-2499098"/>
-              <model>
-                <item value="All"/>
-                <item value="username1"/>
-                <item value="username2"/>
-                <item value="username3"/>
-              </model>
-              <selectionMode value="0"/>
-            </properties>
-          </component>
-        </children>
-      </grid>
-    </children>
-  </grid>
-</form>
